game 2001 a1 v1.2
This is the second commit, this zip has all files in one. This commit adds in the two child classes for the base array class. An ordered list and an unordered list. The ordered list has the option to not allow duplicates and auto sorts upon pushing. The unordered list just acts as a list.